+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 - **irmin-pack**
   - Fix issue when migrating v2 stores to use lower layer (@metanivek, #2241)
+  - Fix issue when calling GC for a commit in the lower after migration
+    (@metanivek, #2242)
 
 ## 3.7.0 (2023-04-26)